Add subprocess checking
The code has an error, it was assuming that all xdg-open processes have
a subprocess but this was not true all the time.

Subprocess check existance was added to avoid runtime exceptions.

Search of the subprocess id is performed cat /proc/<process_id>/cmd-line
to avoid using another application like grep.

Signed-off-by: Nelson Victor Cruz Hernandez <victor.cruz.isc@gmail.com>